Auto-generate module version numbers
The version numbers that are reported by modules reflect the version of
the upstream driver around the time we last pulled it directly into this
project. Because these numbers are basically static, it isn't terribly
useful when diagnosing problems that depend on knowing the exact driver
that is currently loaded. To work around this, this patch appends a
version number which is automatically regenerated (either from git, or
a distributed "version" file) whenever the driver is recompiled.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>