Auto-generate configure.ac version numbers
One of the steps that must be completed prior to making an input-wacom
release is bumping up the version number in configure.ac. This is an
error-prone process (note our missing "0.24.0" release ;)) which can
be replaced by having configure.ac get the current version from our
"git-version-gen" script. With this patch in place, the final commit
should be tagged and then `./autogen.sh` and finally  `make dist` run.
The last two steps can be run at any time to make properly-versioned
beta tarballs as well.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>